requirements.txt
google-play-scraper
pandas
numpy
groq
sentence-transformers
faiss-cpu
